BUG: Update Mesh IO factory to use static initialization. See #3393
This commit updates the Mesh IO factory to ensure it can be built
as shared library so it is registered only once when linked to both
a main application and application plugin dynamically loaded.

In a nutshell, it updates the code so the itk::MeshIOFactory is
registered only once when statically initialized.

More specifically, it applies the following changes:

 (1) Uses of "RegisterFactoryInternal" to make sure the factory
     is registered once into the TransformIOFactory.
     See 39b4ab3 (BUG: Internal factory must use
     RegisterFactoryInternal method) for more details.

 (2) Introduces ITKIOMesh explicit instantiation where
     templated reader and writer are explicitly instantiated over
     dimensions, types and traits.
     This will allow templated factories to be registered in one
     translation unit and instantiated in an other.
     See http://stackoverflow.com/questions/8024010/why-do-template-class-functions-have-to-be-declared-in-the-same-translation-unit

 (3) Update TestDriver to explicitly register the MeshIO
     factories needed by QuadEdgeMeshFiltering and ITKIOMesh tests.

Note that in a follow-up commit, the MeshIO library will be split into
multiple modules MeshIOBase and one for (or one for each) Mesh IO factories.

Open questions:

(1) It is not clear why on Visual Studio 2008, the export
macro need to be removed from MeshFileReader. Otherwise, it reports the
following link error when building ITKIOMeshTestDriver:

  error C2487: 'itk::ProcessObject::SetOutput' : member of dll interface
  class may not be declared with dll interface
    C:\ath/to\ITKv4\Modules\IO\Mesh\include\itkMeshFileReader.h	77


(2) This commits introduces ITKVoronoi dependency to the ITKIOMesh
module. Moving forward, this could be updated by having every module
explicitly listing instantiations they expect.

Change-Id: I80ea34becae742a63b5c98220b20003303188d9c