COMP: TransformIO: Fix VS build error related to explicit instantiation
This commit fixes a regression introduced in eb2316a (BUG: TransformIO:
Implement explicit template instantiation using "extern".). By removing the
use of custom EXPORT_EXPLICIT macro and relying on EXPORTS macro, the
instantiation will be "exported" "extern" only if the associated header is
included by a source file using the library. It means that if within a given
library multiple files include the header using "extern", the corresponding
intantiation will not be declared as "extern".

Error fixed by this commit:

  itkTransformIOFactory.h(83): warning C4910: 'itk::TransformIOFactoryTemplate' : '__declspec(dllexport)' and 'extern' are incompatible on an explicit instantiation

For sake of consistency, this commit also simplifies the explicit
instantiation associated with itkMetaDataObject. It wasn't causing
build error because the associated was included only once in the ITKCommon
library (only in itkMetaDataObject.cxx).

Change-Id: Ie76720b967d09ebedceecfdd62c6e3a243a1c4b1